Code for a 7 button controller. functions for interpreting the adc value output from the controller.
Updated the serialLog() function to Log() in order to simplify for readability and typing purposes. Changed globally.

Changed compilation settings in platformio.ini in order to control which ports output which device. Created a new environment that complies specific code for specific devices. Tested, and can compile two environments accurately.

Created a compiler.h file, may not be necessary (redundant with config.h)